# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: CloverApp Panamá API
  version: 0.1.0
  description: >
    Stub de referencia. La fuente de verdad interactiva se genera automáticamente
    en /docs (Swagger, vía @nestjs/swagger) al levantar el servidor con `npm run dev:api`.
servers:
  - url: http://localhost:3000/v1
security:
  - bearerAuth: []
paths:
...
  /sale/sales/delete/{id}:
    delete:
      tags: [sales]
      summary: Cancelar venta propia (solo antes del cierre)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Venta cancelada } }
  /sale/admin/delete/{id}:
    delete:
      tags: [sales]
      summary: Cancelación administrativa (auditada, sin restricción de cierre)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Venta cancelada por admin } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT